requirements.txt
braintree==3.31.0
colorama==0.3.7
decorator==4.0.10
defusedxml==0.5.0
Django==1.10.2
django-allauth==0.31.0
django-analytical==2.2.2
django-controlcenter==0.2.1
django-cors-headers==2.2.0
django-crispy-forms==1.6.1
django-filter==1.0.0
django-hitcount==1.2.2
django-pkgconf==0.2.0
django-profiles==0.2
django-registration==2.2
django-tinymce4-lite==1.7.0
django-vote==2.1.7
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
ipython==5.1.0
ipython-genutils==0.1.0
jsmin==2.2.2
oauthlib==2.0.2
pickleshare==0.7.4
Pillow==3.4.1
prompt-toolkit==1.0.8
Pygments==2.1.3
PyJWT==1.6.1
pypng==0.0.18
PyQRCode==1.2.1
python-dateutil==2.6.0
python3-openid==3.1.0
pytz==2017.2
requests==2.12.1
requests-oauthlib==0.8.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
uuid==1.30
virtualenv==16.0.0
votes==1.0.1
wcwidth==0.1.7
win-unicode-console==0.5
